refactor:box 컴포넌트를 CardinalInfo 컴포넌트로 옮김
