+++ src/pages/admin/AdminMember.tsx
@@ -1,10 +1,9 @@
-import Box from '@/components/Admin/Box';
+import CardinalInfo from '@/components/Admin/CardinalInfo';
 import MemberListTable, { Column } from '@/components/Admin/MemberListTable';
 import NavMenu from '@/components/Admin/NavMenu';
 import SearchBar from '@/components/Admin/SearchBar';
 import SelectedTopBar from '@/components/Admin/SelectedTopBar';
 import TopBar from '@/components/Admin/TopBar';
-import { BoxWrapper } from '@/components/Admin/TotalDues';
 import {
   MemberProvider,
   useMemberContext,
@@ -14,48 +13,6 @@ import {
   ContentWrapper,
   Container,
 } from '@/styles/admin/AdminLayout.styled';
-import theme from '@/styles/theme';
-import { styled } from 'styled-components';
-
-const CardinalBoxWrapper = styled(BoxWrapper)`
-  padding: 30px 0 30px 0;
-`;
-const boxData = [
-  {
-    id: 'cardinal-total',
-    description: '전체',
-    last: '총 100명',
-    color: `${theme.color.gray[18]}`,
-  },
-  {
-    id: 'cardinal-4',
-    title: '24년 2학기(현재)',
-    description: '4기',
-    last: '동장 노정완 외 25명',
-    color: `${theme.color.gray[65]}`,
-  },
-  {
-    id: 'cardinal-3',
-    title: '24년 1학기',
-    description: '3기',
-    last: '동장 김성민 외 25명',
-    color: `${theme.color.gray[65]}`,
-  },
-  {
-    id: 'cardinal-2',
-    title: '23년 2학기',
-    description: '2기',
-    last: '동장 김성민 외 25명',
-    color: `${theme.color.gray[65]}`,
-  },
-  {
-    id: 'cardinal-1',
-    title: '23년 1학기',
-    description: '1기',
-    last: '동장 김성민 외 25명',
-    color: `${theme.color.gray[65]}`,
-  },
-];
 
 const columns: Column[] = [
   { key: 'name', header: '이름' },
@@ -93,18 +50,7 @@ const AdminMember: React.FC = () => {
           <DynamicTopBar />
           <Container>
             <SearchBar />
-            <CardinalBoxWrapper>
-              {boxData.map((cardinalBox) => (
-                <Box
-                  key={cardinalBox.id}
-                  title={cardinalBox.title}
-                  description={cardinalBox.description}
-                  last={cardinalBox.last}
-                  color={cardinalBox.color}
-                  lastColor="#D3D3D3"
-                />
-              ))}
-            </CardinalBoxWrapper>
+            <CardinalInfo />
             <MemberListTable columns={columns} />
           </Container>
         </ContentWrapper>
